// src/app.types.ts
export type Gender = "Male" | "Female" | "Genderless" | "unknown";
export type Status = "Alive" | "Dead" | "unknown";
export type Species =
  | "Human"
  | "Alien"
  | "Humanoid"
  | "Poopybutthole"
  | "Mythological"
  | "Unknown"
  | "Animal"
  | "Disease"
  | "Robot"
  | "Cronenberg"
  | "Planet";